Subject: integrate and refactor auth module tests (US 08.001 & 08.002)
Body:
- Added missing Sign in test case according to US 08.001.
- Unified Sign in and Sign out tests into a single module.
- Refactored naming conventions to match project documentation.
- Removed time.sleep() calls in favor of WebDriverWait for better stability.
- Implemented pytest-dependency markers to link test execution flow.
